[AIRFLOW-5917][part of AIRFLOW-5893] Group tests for the Webserver command (#6566)
