+++ tests/core.py
@@ -1567,82 +1567,6 @@ class TestCli(unittest.TestCase):
             stdout = mock_stdout.getvalue()
         self.assertIn(version, stdout)
 
-    def _wait_pidfile(self, pidfile):
-        while True:
-            try:
-                with open(pidfile) as file:
-                    return int(file.read())
-            except Exception:
-                sleep(1)
-
-    def test_cli_webserver_foreground(self):
-        # Confirm that webserver hasn't been launched.
-        # pgrep returns exit status 1 if no process matched.
-        self.assertEqual(1, subprocess.Popen(["pgrep", "-c", "airflow"]).wait())
-        self.assertEqual(1, subprocess.Popen(["pgrep", "-c", "gunicorn"]).wait())
-
-        # Run webserver in foreground and terminate it.
-        p = subprocess.Popen(["airflow", "webserver"])
-        p.terminate()
-        p.wait()
-
-        # Assert that no process remains.
-        self.assertEqual(1, subprocess.Popen(["pgrep", "-c", "airflow"]).wait())
-        self.assertEqual(1, subprocess.Popen(["pgrep", "-c", "gunicorn"]).wait())
-
-    @unittest.skipIf("TRAVIS" in os.environ and bool(os.environ["TRAVIS"]),
-                     "Skipping test due to lack of required file permission")
-    def test_cli_webserver_foreground_with_pid(self):
-        # Run webserver in foreground with --pid option
-        pidfile = tempfile.mkstemp()[1]
-        p = subprocess.Popen(["airflow", "webserver", "--pid", pidfile])
-
-        # Check the file specified by --pid option exists
-        self._wait_pidfile(pidfile)
-
-        # Terminate webserver
-        p.terminate()
-        p.wait()
-
-    @unittest.skipIf("TRAVIS" in os.environ and bool(os.environ["TRAVIS"]),
-                     "Skipping test due to lack of required file permission")
-    def test_cli_webserver_background(self):
-        import psutil
-
-        # Confirm that webserver hasn't been launched.
-        self.assertEqual(1, subprocess.Popen(["pgrep", "-c", "airflow"]).wait())
-        self.assertEqual(1, subprocess.Popen(["pgrep", "-c", "gunicorn"]).wait())
-
-        # Run webserver in background.
-        subprocess.Popen(["airflow", "webserver", "-D"])
-        pidfile = cli.setup_locations("webserver")[0]
-        self._wait_pidfile(pidfile)
-
-        # Assert that gunicorn and its monitor are launched.
-        self.assertEqual(0, subprocess.Popen(["pgrep", "-c", "airflow"]).wait())
-        self.assertEqual(0, subprocess.Popen(["pgrep", "-c", "gunicorn"]).wait())
-
-        # Terminate monitor process.
-        pidfile = cli.setup_locations("webserver-monitor")[0]
-        pid = self._wait_pidfile(pidfile)
-        p = psutil.Process(pid)
-        p.terminate()
-        p.wait()
-
-        # Assert that no process remains.
-        self.assertEqual(1, subprocess.Popen(["pgrep", "-c", "airflow"]).wait())
-        self.assertEqual(1, subprocess.Popen(["pgrep", "-c", "gunicorn"]).wait())
-
-    # Patch for causing webserver timeout
-    @mock.patch("airflow.bin.cli.get_num_workers_running", return_value=0)
-    def test_cli_webserver_shutdown_when_gunicorn_master_is_killed(self, _):
-        # Shorten timeout so that this test doesn't take too long time
-        args = self.parser.parse_args(['webserver'])
-        with conf_vars({('webserver', 'web_server_master_timeout'): '10'}):
-            with self.assertRaises(SystemExit) as e:
-                cli.webserver(args)
-        self.assertEqual(e.exception.code, 1)
-
 
 class TestConnection(unittest.TestCase):
     def setUp(self):
